# lib/api-zod/src/generated/types/adminMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for FECh Plenario
 * OpenAPI spec version: 0.1.0
 */
import type { AdminMessageKind } from './adminMessageKind';
import type { AdminMessageRecipient } from './adminMessageRecipient';

export interface AdminMessage {
  id: number;
  subject: string;
  body: string;
  kind: AdminMessageKind;
  /** @nullable */
  replyDeadline?: string | null;
  createdAt: string;
  /** @nullable */
  sessionId?: number | null;
  /** @nullable */
  sessionTitle?: string | null;
  recipients?: AdminMessageRecipient[];
  total?: number;
  leidos?: number;
  confirmadas?: number;
  justificadas?: number;
  sinResponder?: number;
  pendientesRevision?: number;
  /** Ponderación comprometida por quienes confirmaron. Es una proyección para estimar el quórum antes de la sesión, nunca asistencia real.
   */
  pesoConfirmado?: number;
}