# lib/api-zod/src/generated/types/messageReplyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for FECh Plenario
 * OpenAPI spec version: 0.1.0
 */
import type { MessageReplyInputReply } from './messageReplyInputReply';

export interface MessageReplyInput {
  reply: MessageReplyInputReply;
  /** Obligatorio al justificar. */
  reason?: string;
}